criado layout separados par login, e admin, reornanizado a estrutura de pastas da views de admin, feita coreção de dados de controler e rotas para direcionar perfis de login para as rotas corretas e removida tabela do dashbord para listagem no arquivo index para que dashbord receba outra configuração

+++ config/routes.php
@@ -1,10 +1,9 @@
 <?php
 
-use App\Controllers\HomeController;
 use App\Controllers\AuthController;
-use App\Controllers\BuildingController;
+use App\Controllers\Admin\HomeController;
+use App\Controllers\Admin\BuildingController;
 use Core\Router\Route;
-use App\Controllers\TemporariaController;
 
 Route::get('/', [HomeController::class, 'index'])->name('home');
 
@@ -18,27 +17,26 @@ Route::middleware('auth')->group(function () {
 
     // Admin Routes
     Route::middleware('admin')->group(function () {
-        Route::get('/admin', [HomeController::class, 'dashboardAdmin'])->name('dashboard.admin');
+        Route::get('/admin', [HomeController::class, 'dashboard'])->name('admin.dashboard');
 
         // Buildings ----------------------------------------------------------------------------
-        Route::get('/buildings', [BuildingController::class, 'index'])->name('buildings.index');
+        Route::get('/admin/buildings', [BuildingController::class, 'index'])->name('admin.buildings.index');
 
         // Create Building
-        //Route::get('/buildings/new', [ProblemsController::class, 'new'])->name('problems.new');
-        Route::post('/buildings', [BuildingController::class, 'create'])->name('buildings.create');
+        Route::get('/admin/buildings/new', [BuildingController::class, 'new'])->name('admin.buildings.new'); // rota nova
+        Route::post('/admin/buildings', [BuildingController::class, 'create'])->name('admin.buildings.create');
 
         // Retrieve Building
-        Route::get('/buildings/new', [BuildingController::class, 'new'])->name('buildings.new'); // rota nova
-        Route::get('/buildings/page/{page}', [BuildingController::class, 'index'])->name('buildings.paginate');
-        Route::get('/buildings/{id}', [BuildingController::class, 'show'])->name('buildings.show'); // usar essa 
+        Route::get('/admin/buildings/page/{page}', [BuildingController::class, 'index'])->name('admin.buildings.paginate');
+        Route::get('/admin/buildings/{id}', [BuildingController::class, 'show'])->name('admin.buildings.show'); // usar essa 
 
         // Update Building
-        Route::get('/buildings/buildings.list', [BuildingController::class, 'new'])->name('buildings.new'); // rota nova
-        Route::get('/buildings/{id}/edit', [BuildingController::class, 'edit'])->name('buildings.edit');
-        Route::put('/buildings/{id}', [BuildingController::class, 'update'])->name('buildings.update');
+        Route::get('/admin/buildings/buildings.list', [BuildingController::class, 'new'])->name('admin.buildings.new'); // rota nova
+        Route::get('/admin/buildings/{id}/edit', [BuildingController::class, 'edit'])->name('admin.buildings.edit');
+        Route::put('/admin/buildings/{id}', [BuildingController::class, 'update'])->name('admin.buildings.update');
 
         // Delete Building
-        Route::delete('/buildings/{id}', [BuildingController::class, 'destroy'])->name('buildings.destroy');
+        Route::delete('/admin/buildings/{id}', [BuildingController::class, 'destroy'])->name('admin.buildings.destroy');
     });
 
     // User Routes
@@ -46,10 +44,3 @@ Route::middleware('auth')->group(function () {
         Route::get('/client', [HomeController::class, 'dashboardClient'])->name('dashboard.client');
     });
 });
-
-// Rotas temporária para teste do componente
-Route::get('/components/navbar.admin', [TemporariaController::class, 'navbarADM']);
-Route::get('/components/sidenav.admin', [TemporariaController::class, 'sidenavADM']);
-Route::get('/components/navbar.user', [TemporariaController::class, 'navbarUser']);
-Route::get('/components/sidenav.user', [TemporariaController::class, 'sidebarUser']);
-//Route::get('/buildings/edit', [TemporariaController::class, 'edit']);
